feat: (wip) refactor EditorPanel to use EditorResources component for resource loading
